requirements.txt
# Core Dependencies
pyyaml>=6.0
python-dotenv>=0.20.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# HTTP & Network
requests>=2.28.0
urllib3>=1.26.0

# Type Checking (Optional)
# mypy>=0.990
# typing-extensions>=4.0.0

# Future: Web Framework
# Flask>=2.3.0
# flask-cors>=4.0.0

# Future: Database
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0

# Future: Scanning Tools
# nmap-python  (if available)
# scapy>=2.5.0
# paramiko>=2.12.0

# Development
black>=22.0.0
flake8>=5.0.0
isort>=5.11.0

